Avoid memoizing arguments of coumpound terms that are already memoized.
Fixed a memory problem due to duplicated entries in the store.

The problem and its solution are demonstrated by the benchmark

  shared-tree.hs

that creates a binary tree whose subtrees are identical.

In this example with a lot of sharing we do not outperform MCC.
MCC is about 10 times slower than a deterministic Haskell program
and about 2.5 times faster than a monadic Haskell program.

Hence, the slowdown compared to a deterministic program is about 25.